Add(jiwon/image): s3를 이용한 이미지 처리 ts 변환 #55
