--- app/src/apis/image/index.js
@@ -1,8 +1,0 @@
-const s3 = require("../../middlewares/s3");
-const express = require("express");
-const router = express.Router();
-const ctrl = require("./image.ctrl");
-
-router.post("/", s3.upload.array("upload", 5), ctrl.process.upload);
-router.delete("/", ctrl.process.delete);
-module.exports = router;
